+++ src/scruby/constants.py
@@ -3,7 +3,7 @@
 The module contains the following variables:
 
 - `DB_ROOT` - Path to root directory of database. `By default = "ScrubyDB"` (*in root of project*).
-- `LENGTH_SEPARATED_HASH` - Length of separated hash for create path inside collection.
+- `LENGTH_REDUCTION_HASH` - The length of the hash reduction on the left side.
     - `0` - 4294967296 keys (by default).
     - `2` - 16777216 keys.
     - `4` - 65536 keys.
@@ -14,7 +14,7 @@ from __future__ import annotations
 
 __all__ = (
     "DB_ROOT",
-    "LENGTH_SEPARATED_HASH",
+    "LENGTH_REDUCTION_HASH",
 )
 
 from typing import Literal
@@ -23,9 +23,9 @@ from typing import Literal
 # By default = "ScrubyDB" (in root of project).
 DB_ROOT: str = "ScrubyDB"
 
-# Length of separated hash for create path inside collection.
+# The length of the hash reduction on the left side.
 # 0 = 4294967296 keys (by default).
 # 2 = 16777216 keys.
 # 4 = 65536 keys.
 # 6 = 256 keys (main purpose is tests).
-LENGTH_SEPARATED_HASH: Literal[0, 2, 4, 6] = 0
+LENGTH_REDUCTION_HASH: Literal[0, 2, 4, 6] = 0